requirements.txt
fastapi
uvicorn
Pillow>=10.0
